+++ src/enums/index.ts
@@ -1,3 +1,4 @@
 export { ROUTE_NAMES } from './route-names.enum'
 export { ICON_NAMES } from './icon-names.enum'
+export { REQUEST_STATUSES } from '@/enums/statuses.enum'
 export * from './window-breakpoints.enum'