+++ src/enums/statuses.enum.ts
@@ -1,0 +1,7 @@
+export enum REQUEST_STATUSES {
+  all = 'all',
+  failed = 'failed',
+  pending = 'pending',
+  created = 'created',
+  finished = 'finished',
+}